+++ src/commands/Dashboard.ts
@@ -434,7 +434,7 @@ export class Dashboard {
     if (crntSort?.type === SortType.string) {
       allMedia = allMedia.sort(Sorting.alphabetically("fsPath"));
     } else if (crntSort?.type === SortType.date) {
-      allMedia = allMedia.sort(Sorting.date("mtime"));
+      allMedia = allMedia.sort(Sorting.dateWithFallback("mtime", "fsPath"));
     } else {
       allMedia = allMedia.sort(Sorting.alphabetically("fsPath"));
     }
@@ -548,7 +548,7 @@ export class Dashboard {
     if (folderInfo) {
       for (const folder of folderInfo) {
         for (const file of folder.lastModified) {
-          if (file.fileName.endsWith(`.md`) || file.fileName.endsWith(`.mdx`)) {
+          if (file.fileName.endsWith(`.md`) || file.fileName.endsWith(`.markdown`) || file.fileName.endsWith(`.mdx`)) {
             try {
               const article = ArticleHelper.getFrontMatterByPath(file.filePath);