+++ src/helpers/Sorting.ts
@@ -35,6 +35,31 @@ export class Sorting {
   };
 
   /**
+   * Sort by date with a fallback
+   * @param property 
+   * @returns 
+   */
+  public static dateWithFallback = (property: string, fallback: string) => {
+    return (a: any, b: any) => {
+      const dateA = DateHelper.tryParse(a[property]);
+      const dateB = DateHelper.tryParse(b[property]);
+
+      // Sort by date
+      var dCount = (dateA || new Date(0)).getTime() - (dateB || new Date(0)).getTime();
+      if(dCount) return dCount;
+
+      // If there is a tie, sort by fallback property
+      if (a[fallback] < b[fallback]) {
+        return -1;
+      }
+      if (a[fallback] > b[fallback]) {
+        return 1;
+      }
+      return 0;
+    };
+  };
+
+  /**
    * Sort by number
    * @param property 
    * @returns